make pytest pickup all the tests, deactivate broken ones
