--- src/tests/power_flow_test.py
@@ -1,31 +1,0 @@
-from GridCal.Engine.IO.file_handler import *
-from GridCal.Engine.Simulations.PowerFlow.power_flow_driver import PowerFlowOptions, ReactivePowerControlMode, PowerFlow, SolverType
-
-if __name__ == '__main__':
-
-    fname = os.path.join('..', '..', 'Grids_and_profiles', 'grids', 'IEEE 30 Bus with storage.xlsx')
-
-    print('Reading...')
-    main_circuit = FileOpen(fname).open()
-    options = PowerFlowOptions(SolverType.NR, verbose=False,
-                               initialize_with_existing_solution=False,
-                               multi_core=False, dispatch_storage=True,
-                               control_q=ReactivePowerControlMode.NoControl,
-                               control_p=True)
-
-    # grid.export_profiles('ppppppprrrrroooofiles.xlsx')
-    # exit()
-
-    ####################################################################################################################
-    # PowerFlow
-    ####################################################################################################################
-    print('\n\n')
-    power_flow = PowerFlow(main_circuit, options)
-    power_flow.run()
-
-    print('\n\n', main_circuit.name)
-    print('\t|V|:', abs(power_flow.results.voltage))
-    print('\t|Sbranch|:', abs(power_flow.results.Sbranch))
-    print('\t|loading|:', abs(power_flow.results.loading) * 100)
-    print('\tReport')
-    print(power_flow.results.get_report_dataframe())
