+++ src/tests/test_branch_tolerance.py
@@ -1,4 +1,11 @@
-from GridCal.Engine import *
+from GridCal.Engine.Core.multi_circuit import MultiCircuit
+from GridCal.Engine.Devices.branch import Branch
+from GridCal.Engine.Devices.bus import Bus
+from GridCal.Engine.Devices.generator import Generator
+from GridCal.Engine.Devices.load import Load
+from GridCal.Engine.Simulations.PowerFlow.power_flow_driver import \
+    PowerFlowOptions, PowerFlow
+from GridCal.Engine.basic_structures import BranchImpedanceMode
 
 Sbase = 100  # MVA
