+++ src/tests/test_reverse_transformer_definition.py
@@ -1,59 +1,27 @@
-from GridCal.Engine import *
 from math import sqrt
 
+from GridCal.Engine.Devices.transformer import TransformerType
 
-if __name__ == '__main__':
 
-    Vf = 20.0  # kV
-    Vt = 0.4  # kV
-    Sn = 0.5  # MVA
-    Vsc_ = 6.0  # %
-    Pcu_ = 6.0  # kW
-    # Pfe_ = 1.4  # kW
-    # I0_ = 0.28  # %
-    Pfe_ = 0.0  # kW
-    I0_ = 0.0  # %
-
-    tpe = TransformerType(hv_nominal_voltage=Vf,
-                          lv_nominal_voltage=Vt,
-                          nominal_power=Sn,
-                          copper_losses=Pcu_,
-                          iron_losses=Pfe_,
-                          no_load_current=I0_,
-                          short_circuit_voltage=Vsc_,
-                          gr_hv1=0.5,
-                          gx_hv1=0.5)
-
-    z, zl = tpe.get_impedances()
-    print(z)
-    print(zl)
-
+def test_reverse_transformer():
     # ------------------------------------------------------------------------------------------------------------------
     # Revert the calcs
     # ------------------------------------------------------------------------------------------------------------------
-    if zl.real > 0 and zl.imag > 0:
-        yl = 1.0 / zl
-        G = yl.real
-        B = yl.imag
-    else:
-        G = 0
-        B = 0
-
-    R = z.real
-    X = z.imag
-
-    Sn = Sn
-
+    Vf = 11
+    Vt = 132
+    G = 0
+    B = 0
+    R = 0
+    X = 0.115
+    Sn = 30
     print()
     print('R', R)
     print('X', X)
     print('G', G)
     print('B', B)
-
     zsc = sqrt(R * R + 1 / (X * X))
     Vsc = 100.0 * zsc
     Pcu = R * Sn * 1000.0
-
     if abs(G) > 0.0 and abs(B) > 0.0:
         zl = 1.0 / complex(G, B)
         rfe = zl.real
@@ -64,14 +32,12 @@ if __name__ == '__main__':
         k = 1 / (rfe * rfe) + 1 / (xm * xm)
         I0 = 100.0 * sqrt(k)
     else:
-        Pfe = 1e20
-        I0 = 1e20
-
-    print('Vsc', Vsc, Vsc_)
-    print('Pcu', Pcu, Pcu_)
-    print('I0', I0, I0_)
-    print('Pfe', Pfe, Pfe_)
-
+        Pfe = 1e-20
+        I0 = 1e-20
+    print('Vsc', Vsc)
+    print('Pcu', Pcu)
+    print('I0', I0)
+    print('Pfe', Pfe)
     tpe2 = TransformerType(hv_nominal_voltage=Vf,
                            lv_nominal_voltage=Vt,
                            nominal_power=Sn,
@@ -81,7 +47,16 @@ if __name__ == '__main__':
                            short_circuit_voltage=Vsc,
                            gr_hv1=0.5,
                            gx_hv1=0.5)
-
     z2, zl2 = tpe2.get_impedances()
-    print(z2)
-    print(zl2)
+    # print(z2)
+    # print(1/zl2)
+    yl = 1 / zl2
+    print()
+    print('R', z2.real)
+    print('X', z2.imag)
+    print('G', yl.real)
+    print('B', yl.imag)
+
+
+if __name__ == '__main__':
+    test_reverse_transformer()
