+++ src/tests/test_temp_correction.py
@@ -1,6 +1,12 @@
-from GridCal.Engine import *
-
 from pytest import approx
+
+from GridCal.Engine.Core.multi_circuit import MultiCircuit
+from GridCal.Engine.Devices.branch import Branch
+from GridCal.Engine.Devices.bus import Bus
+from GridCal.Engine.Devices.generator import Generator
+from GridCal.Engine.Devices.load import Load
+from GridCal.Engine.Simulations.PowerFlow.power_flow_driver import \
+    PowerFlowOptions, PowerFlow
 
 Sbase = 100  # MVA
