+++ src/tests/test_transformer_regulator.py
@@ -1,4 +1,13 @@
-from GridCal.Engine import *
+from GridCal.Engine.Core.multi_circuit import MultiCircuit
+from GridCal.Engine.Devices.branch import Branch, TapChanger
+from GridCal.Engine.Devices.bus import Bus
+from GridCal.Engine.Devices.generator import Generator
+from GridCal.Engine.Devices.static_generator import StaticGenerator
+from GridCal.Engine.Devices.transformer import TransformerType
+from GridCal.Engine.Devices.types import BranchType
+from GridCal.Engine.Simulations.PowerFlow.power_flow_driver import \
+    PowerFlowOptions, SolverType, ReactivePowerControlMode, TapsControlMode, \
+    PowerFlowMP
 
 Sbase = 100  # MVA
