+++ src/tests/test_voltage_controlled_generator.py
@@ -1,6 +1,12 @@
-from GridCal.Engine import *
-
-Sbase = 100 # MVA
+from GridCal.Engine.Core.multi_circuit import MultiCircuit
+from GridCal.Engine.Devices.branch import Branch, TapChanger
+from GridCal.Engine.Devices.bus import Bus
+from GridCal.Engine.Devices.generator import Generator
+from GridCal.Engine.Devices.transformer import TransformerType
+from GridCal.Engine.Devices.types import BranchType
+from GridCal.Engine.Simulations.PowerFlow.power_flow_driver import \
+    PowerFlowOptions, PowerFlow, SolverType, ReactivePowerControlMode, \
+    TapsControlMode
 
 
 def complex_impedance(z, XR):
@@ -9,9 +15,9 @@ def complex_impedance(z, XR):
     """
     z = float(abs(z))
     XR = float(abs(XR))
-    real = (z**2/(1+XR**2))**0.5
+    real = (z ** 2 / (1 + XR ** 2)) ** 0.5
     try:
-        imag = (z**2/(1+1/XR**2))**0.5
+        imag = (z ** 2 / (1 + 1 / XR ** 2)) ** 0.5
     except ZeroDivisionError:
         imag = 0.0
     return complex(real, imag)
@@ -28,22 +34,23 @@ def test_pv_1():
     """
     test_name = "test_pv_1"
     grid = MultiCircuit(name=test_name)
+    Sbase = 100  # MVA
     grid.Sbase = Sbase
     grid.time_profile = None
     grid.logger = list()
 
     # Create buses
     POI = Bus(name="POI",
-              vnom=100, #kV
+              vnom=100,  # kV
               is_slack=True)
     grid.add_bus(POI)
 
     B_MV_M32 = Bus(name="B_MV_M32",
-                   vnom=10) #kV
+                   vnom=10)  # kV
     grid.add_bus(B_MV_M32)
 
     B_LV_M32 = Bus(name="B_LV_M32",
-                   vnom=0.6) #kV
+                   vnom=0.6)  # kV
     grid.add_bus(B_LV_M32)
 
     # Create voltage controlled generators (or slack, a.k.a. swing)
@@ -60,29 +67,31 @@ def test_pv_1():
     grid.add_generator(B_LV_M32, M32)
 
     # Create transformer types
-    s = 100 # MVA
-    z = 8 # %
+    s = 100  # MVA
+    z = 8  # %
     xr = 40
     SS = TransformerType(name="SS",
-                         hv_nominal_voltage=100, # kV
-                         lv_nominal_voltage=10, # kV
+                         hv_nominal_voltage=100,  # kV
+                         lv_nominal_voltage=10,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=125, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=125,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(SS)
 
-    s = 5 # MVA
-    z = 6 # %
+    s = 5  # MVA
+    z = 6  # %
     xr = 20
     PM = TransformerType(name="PM",
-                         hv_nominal_voltage=10, # kV
-                         lv_nominal_voltage=0.6, # kV
+                         hv_nominal_voltage=10,  # kV
+                         lv_nominal_voltage=0.6,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=6.25, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=6.25,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(PM)
 
@@ -120,8 +129,9 @@ def test_pv_1():
     power_flow = PowerFlow(grid, options)
     power_flow.run()
 
-    approx_volt = [round(100*abs(v), 1) for v in power_flow.results.voltage]
-    solution = [100.0, 100.1, 102.5] # Expected solution from GridCal and ETAP 16.1.0, for reference
+    approx_volt = [round(100 * abs(v), 1) for v in power_flow.results.voltage]
+    solution = [100.0, 100.1,
+                102.5]  # Expected solution from GridCal and ETAP 16.1.0, for reference
 
     print()
     print(f"Test: {test_name}")
@@ -139,19 +149,21 @@ def test_pv_1():
         print(f" - {b}:")
         print(f"   R = {round(b.R, 4)} pu")
         print(f"   X = {round(b.X, 4)} pu")
-        print(f"   X/R = {round(b.X/b.R, 1)}")
+        print(f"   X/R = {round(b.X / b.R, 1)}")
         print(f"   G = {round(b.G, 4)} pu")
         print(f"   B = {round(b.B, 4)} pu")
     print()
 
     print("Transformer types:")
     for t in grid.transformer_types:
-        print(f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
+        print(
+            f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
     print()
 
     print("Losses:")
     for i in range(len(grid.branches)):
-        print(f" - {grid.branches[i]}: losses={1000*round(power_flow.results.losses[i], 3)} kVA")
+        print(
+            f" - {grid.branches[i]}: losses={1000 * round(power_flow.results.losses[i], 3)} kVA")
     print()
 
     equal = True
@@ -174,22 +186,23 @@ def test_pv_2():
     """
     test_name = "test_pv_2"
     grid = MultiCircuit(name=test_name)
+    Sbase = 100  # MVA
     grid.Sbase = Sbase
     grid.time_profile = None
     grid.logger = list()
 
     # Create buses
     POI = Bus(name="POI",
-              vnom=100, #kV
+              vnom=100,  # kV
               is_slack=True)
     grid.add_bus(POI)
 
     B_MV_M32 = Bus(name="B_MV_M32",
-                   vnom=10) #kV
+                   vnom=10)  # kV
     grid.add_bus(B_MV_M32)
 
     B_LV_M32 = Bus(name="B_LV_M32",
-                   vnom=0.6) #kV
+                   vnom=0.6)  # kV
     grid.add_bus(B_LV_M32)
 
     # Create voltage controlled generators (or slack, a.k.a. swing)
@@ -206,29 +219,31 @@ def test_pv_2():
     grid.add_generator(B_LV_M32, M32)
 
     # Create transformer types
-    s = 100 # MVA
-    z = 8 # %
+    s = 100  # MVA
+    z = 8  # %
     xr = 40
     SS = TransformerType(name="SS",
-                         hv_nominal_voltage=100, # kV
-                         lv_nominal_voltage=10, # kV
+                         hv_nominal_voltage=100,  # kV
+                         lv_nominal_voltage=10,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=125, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=125,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(SS)
 
-    s = 5 # MVA
-    z = 6 # %
+    s = 5  # MVA
+    z = 6  # %
     xr = 20
     PM = TransformerType(name="PM",
-                         hv_nominal_voltage=10, # kV
-                         lv_nominal_voltage=0.6, # kV
+                         hv_nominal_voltage=10,  # kV
+                         lv_nominal_voltage=0.6,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=6.25, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=6.25,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(PM)
 
@@ -240,7 +255,8 @@ def test_pv_2():
                   template=SS,
                   bus_to_regulated=True,
                   vset=1.005)
-    X_C3.tap_changer = TapChanger(taps_up=16, taps_down=16, max_reg=1.1, min_reg=0.9)
+    X_C3.tap_changer = TapChanger(taps_up=16, taps_down=16, max_reg=1.1,
+                                  min_reg=0.9)
     X_C3.tap_changer.set_tap(X_C3.tap_module)
     grid.add_branch(X_C3)
 
@@ -271,8 +287,8 @@ def test_pv_2():
     power_flow = PowerFlow(grid, options)
     power_flow.run()
 
-    approx_volt = [round(100*abs(v), 1) for v in power_flow.results.voltage]
-    solution = [100.0, 100.7, 102.5] # Expected solution from GridCal
+    approx_volt = [round(100 * abs(v), 1) for v in power_flow.results.voltage]
+    solution = [100.0, 100.7, 102.5]  # Expected solution from GridCal
 
     print()
     print(f"Test: {test_name}")
@@ -290,19 +306,21 @@ def test_pv_2():
         print(f" - {b}:")
         print(f"   R = {round(b.R, 4)} pu")
         print(f"   X = {round(b.X, 4)} pu")
-        print(f"   X/R = {round(b.X/b.R, 1)}")
+        print(f"   X/R = {round(b.X / b.R, 1)}")
         print(f"   G = {round(b.G, 4)} pu")
         print(f"   B = {round(b.B, 4)} pu")
     print()
 
     print("Transformer types:")
     for t in grid.transformer_types:
-        print(f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
+        print(
+            f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
     print()
 
     print("Losses:")
     for i in range(len(grid.branches)):
-        print(f" - {grid.branches[i]}: losses={1000*round(power_flow.results.losses[i], 3)} kVA")
+        print(
+            f" - {grid.branches[i]}: losses={1000 * round(power_flow.results.losses[i], 3)} kVA")
     print()
 
     equal = True
@@ -328,22 +346,23 @@ def test_pv_3():
     """
     test_name = "test_pv_3"
     grid = MultiCircuit(name=test_name)
+    Sbase = 100  # MVA
     grid.Sbase = Sbase
     grid.time_profile = None
     grid.logger = list()
 
     # Create buses
     POI = Bus(name="POI",
-              vnom=100, #kV
+              vnom=100,  # kV
               is_slack=True)
     grid.add_bus(POI)
 
     B_MV_M32 = Bus(name="B_MV_M32",
-                   vnom=10) #kV
+                   vnom=10)  # kV
     grid.add_bus(B_MV_M32)
 
     B_LV_M32 = Bus(name="B_LV_M32",
-                   vnom=0.6) #kV
+                   vnom=0.6)  # kV
     grid.add_bus(B_LV_M32)
 
     # Create voltage controlled generators (or slack, a.k.a. swing)
@@ -360,29 +379,31 @@ def test_pv_3():
     grid.add_generator(B_LV_M32, M32)
 
     # Create transformer types
-    s = 100 # MVA
-    z = 8 # %
+    s = 100  # MVA
+    z = 8  # %
     xr = 40
     SS = TransformerType(name="SS",
-                         hv_nominal_voltage=100, # kV
-                         lv_nominal_voltage=10, # kV
+                         hv_nominal_voltage=100,  # kV
+                         lv_nominal_voltage=10,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=125, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=125,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(SS)
 
-    s = 5 # MVA
-    z = 6 # %
+    s = 5  # MVA
+    z = 6  # %
     xr = 20
     PM = TransformerType(name="PM",
-                         hv_nominal_voltage=10, # kV
-                         lv_nominal_voltage=0.6, # kV
+                         hv_nominal_voltage=10,  # kV
+                         lv_nominal_voltage=0.6,  # kV
                          nominal_power=s,
-                         copper_losses=complex_impedance(z, xr).real*s*1000/Sbase,
-                         iron_losses=6.25, # kW
-                         no_load_current=0.5, # %
+                         copper_losses=complex_impedance(z,
+                                                         xr).real * s * 1000 / Sbase,
+                         iron_losses=6.25,  # kW
+                         no_load_current=0.5,  # %
                          short_circuit_voltage=z)
     grid.add_transformer_type(PM)
 
@@ -394,7 +415,8 @@ def test_pv_3():
                   template=SS,
                   bus_to_regulated=True,
                   vset=1.005)
-    X_C3.tap_changer = TapChanger(taps_up=16, taps_down=16, max_reg=1.1, min_reg=0.9)
+    X_C3.tap_changer = TapChanger(taps_up=16, taps_down=16, max_reg=1.1,
+                                  min_reg=0.9)
     X_C3.tap_changer.set_tap(X_C3.tap_module)
     grid.add_branch(X_C3)
 
@@ -425,8 +447,8 @@ def test_pv_3():
     power_flow = PowerFlow(grid, options)
     power_flow.run()
 
-    approx_volt = [round(100*abs(v), 1) for v in power_flow.results.voltage]
-    solution = [100.0, 100.7, 102.5] # Expected solution from GridCal
+    approx_volt = [round(100 * abs(v), 1) for v in power_flow.results.voltage]
+    solution = [100.0, 100.7, 102.5]  # Expected solution from GridCal
 
     print()
     print(f"Test: {test_name}")
@@ -444,19 +466,21 @@ def test_pv_3():
         print(f" - {b}:")
         print(f"   R = {round(b.R, 4)} pu")
         print(f"   X = {round(b.X, 4)} pu")
-        print(f"   X/R = {round(b.X/b.R, 1)}")
+        print(f"   X/R = {round(b.X / b.R, 1)}")
         print(f"   G = {round(b.G, 4)} pu")
         print(f"   B = {round(b.B, 4)} pu")
     print()
 
     print("Transformer types:")
     for t in grid.transformer_types:
-        print(f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
+        print(
+            f" - {t}: Copper losses={int(t.Pcu)}kW, Iron losses={int(t.Pfe)}kW, SC voltage={t.Vsc}%")
     print()
 
     print("Losses:")
     for i in range(len(grid.branches)):
-        print(f" - {grid.branches[i]}: losses={1000*round(power_flow.results.losses[i], 3)} kVA")
+        print(
+            f" - {grid.branches[i]}: losses={1000 * round(power_flow.results.losses[i], 3)} kVA")
     print()
 
     equal = True
